+++ NEWS.md
@@ -3,7 +3,7 @@
 # poiscon 0.6.6
 
 * added `install_hd` function to simplify installing packages from harddrive
-* added `git_commit` function to commit (and push) current (or other git repository)
+* added internal `git_commit` function to commit (and push) for current (or other git repository)
 
 # poiscon 0.6.5